api/requirements.txt
python-dotenv==1.1.1
fastapi==0.116.1
uvicorn==0.35.0
boto3==1.40.33
python-multipart==0.0.20
httpx==0.27.2
requests==2.32.3
aiohttp==3.10.5
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow>=9.0.0
